+++ src/background/session/type.ts
@@ -1,3 +1,0 @@
-export interface SessionState {
-  hasBrowserActionMenu?: boolean;
-}